Improved error handling if board is set up badly
The light sensor error message is now printed before the board ID, so that it causes an error.  This could probably be done more elegantly, but it's ok for now.

I've also added error handling to the constructor so it doesn't leave the port open in the event that we can't initialise the board.