Add test mode
Stage now has a test_mode property which is propagated to modules and the firmware.
With test_mode on the stage will return more data about the homing procedure, will
stay at 0 after homing and will not reset position when it hits an endstop.

This property may be used in the future to report more data/change other behaviour
as required for testing.